feat(upgrader-security): Validate and publish the signed manifest (#1494)
#### Details

This PR adds 2 things:
1. Include the signed manifest in BinSkim and signing validation (not
strictly necessary since we don't directly ship the assembly, but it
seemed appropriate since we rely on the contents being secured)
2. Publish the manifest as a build artifact so that the release pipeline
can easily consume it

The manifest is published as the `manifest` artifact of this test build
of the signed pipeline:
https://dev.azure.com/mseng/1ES/_build/results?buildId=18867199&view=artifacts&pathAsName=false&type=publishedArtifacts

##### Motivation

Part of the upgrader security work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue# - 
- [n/a] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.