feat(audio-feedback): Introduce the ternary setting (#1495)
In #1445, I mentioned the possibility of switching to a ternary (auto/always/never) setting for sound feedback:

* "automatic" plays sound according to the system screen reader flag (default).
* "always" always plays a scanning sound.
* "never" does not play a sound in any circumstances, equivalent to the previous default behaviour.

This commit replaces the current binary `PlayScanningSound` option with a new ternary setting, mapping `true` to `Always`.

Part of refining the sound feedback strategy (#1440, #1444).

Better selection of "which sound to play" in `ProgressRingControl` to follow in subsequent PRs (I wanted to make the changes here as minimal as possible).

There are some odd behaviours that repro intermittently (scanning sound played for a few seconds after I Alt+Tabbed), and in one case it played infinitely during an update, but I can't consistently and reliably determine that they're caused by this PR (i.e. they might be current issues).