feat(upgrader-security): Add integration tests for manifests (#1505)
#### Details

Add tests to catch potential manifest mismatches at build time. This is
the "future PR" mentioned in the "Context" paragraph of #1493. These
tests essentially execute the code that will be run on the client to
make sure that the manifest is available and that it contains the
expected data. Since these tests rely on the `MSI` and `Manifest`
projects, and since these projects are excluded from debug builds, the
`ManifestTests` project is likewise excluded from debug builds.

##### Motivation

Keep manifest mechanism robust

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
I wanted to add a test that would check for signing of the manifest
assembly, but that wouldn't play very nicely with unsigned builds. The
signed pipeline already verifies the manifest for correct signing, so it
seemed safe to omit that test. I also had to create a way to bypass the
`TrustVerifier` check for unsigned builds. It would be cool if this
bypass were skipped in the `SignedRelease` build, but I wasn't sure how
to make that work.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue# - 
- [n/a] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.