feat: Improve discovery of keyboard shortcuts (#1509)
#### Details

New keyboard users may not be familiar with the metaphor of using F6 to
move between panes. To make this a bit easier to discover, we're
including this information and a link to the full keyboard shortcuts in
the Live Inspect pane. The link goes through a fwlink and resolves to
https://accessibilityinsights.io/docs/windows/reference/keyboard/.

##### Motivation

Improve the UX for new keyboard users.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
Like many other places in the code, this uses string concatenation and
is not really localizable. I tried a version that used a RegEx, but the
Hyperlink styling was wrong. The same problem exists in all 3 text
blocks in this control. so I'm deferring solving that problem for a
later time if/when we take on proper localization.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
We discussed putting this into the Welcome screen, but decided against
it for the following reasons:
1. That dialog is already at its limit at 200% zoom
2. Should the user ever uncheck the "Show this dialog at startup"
checkbox, we don't provide an easy way to get it back

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

#### Screenshots

Before | After
--- | ---

![image](https://user-images.githubusercontent.com/45672944/206061859-3c515e64-ce05-4504-92b6-cf21a2ff60b0.png)
|
![image](https://user-images.githubusercontent.com/45672944/206061669-be9e1a24-2d46-4168-a791-78160f8f0351.png)


- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? 
- [x] Includes UI changes?
- [x] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [x] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.