fix: Use size of and position in set for event config dialog tabs (#1532)
#### Details

Replace hardcoded "x of 3" in element automation name with proper UIA
`SizeOfSet` and `PositionInSet` properties. This PR addresses the 2 tabs
in in the event config dialog, which can be found as follows:

- Open AI win in live view and inspect any app
- Select 'More options" and "Listen to events" in drop downs.
- In right hand side "Configuration" pane, expand the "My Events" item
in the events list
- Select the "Edit my events..." link under "My Events"
- See event configuration dialog pop up. Observe the 2 "Events" and
"Properties" tabs in the dialog

Note that this PR changed the size of set value from 3 to 2. There are
only 2 tabs so I am not sure why this was ever 'x of 3'. If I'm missing
something here, please let me know and I will change it back.

##### Motivation

Addresses part of
https://github.com/microsoft/accessibility-insights-windows/issues/1514

##### Context

This is one of several PRs to address the various issues discussed in
https://github.com/microsoft/accessibility-insights-windows/issues/1514

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [x] Does this address an existing issue? If yes,
https://github.com/microsoft/accessibility-insights-windows/issues/1514
- [n/a] Includes UI changes?
- [ ] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.