feat(upgrader-security): Improve manifest handling (#1558)
#### Details

This improves upgrade manifest _handling_ (content and structure are
unmodified) in 2 ways:
1. Rather than needing to manually edit the value of
`MinimumProdVersion` in
[version.props](https://github.com/microsoft/accessibility-insights-windows/blob/main/src/props/prod-version.props)
after each production release, we now populate the manifest's
`production_minimum_version` field automatically based on the value of
`IsMandatoryProdUpdate`, which will only need to changed if a mandatory
upgrade released (which has happened **exactly once since 2019**):
- If `IsMandatoryProdUpdate` is `false` (the usual state), then
`production_minimum_version` is the same as the latest non-prerelease
release from GitHub. This allows production clients to remain 1 version
behind the latest release when this is released to production,
consistent with our existing policy.
- If `IsMandatoryProdUpdate` is `true`, then
`production_minimum_version` is the same as `current_version`. This
forces all production clients running earlier versions to upgrade when
this is released to production.
2. We will now include the unsigned manifest in the `Manifest` build
artifact. The objective is to replace human-provided data (specified
when creating the Production release) with data from the `Manifest`
build artifact. This reduces the risk of human error at release creation
time.

##### Motivation

Streamline the release process

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
This creates a new potential failure point, since the release build will
fail if the GitHub API is down. Most local builds are debug (which does
_not_ build the manifest), so the risk here is very low. The build
pipeline might break, but if the GitHub API is down, it will probably
break much earlier in the process. Altogether, the risk seems acceptably
low. If we encounter too many build breaks with this new dependency, we
can consider options to provide a workaround, possibly as a pipeline
variable. YAGNI until proven otherwise.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [ ] Does this address an existing issue? If yes, Issue# - 
- [ ] Includes UI changes?
- [ ] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.