feat(upgrader-security): Deprecate unsigned manifests (#1559)
#### Details

Now that all channels are using signed manifests, we can deprecate
client support for unsigned manifests. This PR removes the code and
tests that supported unsigned manifests. It also adds a test to be sure
that we can always read the sign Canary manifest. Strictly speaking,
this test is an integration test instead of a unit test, but the unit
test class is so small that it seemed reasonable to put both types of
tests into the same class.

##### Motivation

Complete upgrader security migration.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [ ] Does this address an existing issue? If yes, Issue# - 
- [ ] Includes UI changes?
- [ ] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.