chore: Revert workaround for broken thirdpartynotice generator (#1560)
#### Details

This reverts the change from #1547, which was added as a temporary
workaround for a bug in the API that generates our
thirdpartynotices.html file. The API has since been fixed so we don't
need the workaround.

##### Motivation

It's preferable to not have the `thirdpartynotices.html` file in the
repo

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [ ] Does this address an existing issue? If yes, Issue# - 
- [ ] Includes UI changes?
- [ ] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.