chore: Move deployment scripts into build artifact (#1577)
#### Details

ADO release pipelines are limited to classic pipeline definitions, with
no YAML support. Our current release pipelines include 5 inline
PowerShell scripts, which are hard to discover and harder to review when
changes are needed. After discussion as a team, we opted to add the
scripts to the repo, publish them as build artifacts, then have the
release pipelines execute the scripts that we consume from the build
artifacts. This PR completes the first 2 steps, then we'll change the
release pipelines once the artifact exists.

Note that while the consumption pattern will change from inline to an
artifact, the scripts themselves will not be changing (caveat: 2
non-functional tweaks were made). The publishing of the `deploy_scripts`
artifact occurs in the SignedRelease step, so it happens at the very end
of the build. It was earlier in a test build just for the sake of
efficiency.

##### Motivation

Take advantage of code tools to manage our release pipelines.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
The pipeline changes will take effect after this is merged and the
`deploy_scripts` artifact becomes available.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [ ] Does this address an existing issue? If yes, Issue# - 
- [ ] Includes UI changes?
- [ ] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.