chore: Increase dependabot PR limit (#1598)
#### Details

Dependabot defaults to a limit of 5 open PR's at a time. The ADO bug
filing code uses 4 assemblies from Visual Studio that all get released
at the same time and that need to merge in a specific order. The last 2
times these got bumped, we ended up with not all of the updates and
needed manual updates to make things work as expected. This PR updates
the dependabot limit to 10, which allows more "breathing room" when
these 4 packages all get updated.

##### Motivation

Make it easier for dependabot to do the grunt work that we want it to do

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [ ] Does this address an existing issue? If yes, Issue# - 
- [ ] Includes UI changes?
- [ ] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.