chore: Fix test irregularities (#1687)
#### Details

I noticed today that the `AppLayoutTests` were not running as expected.
In investigating, I found the following, which are fixed in this PR:
- The `AppLayoutTests.ClassInitialize` method needs to take a
`TestContext` as a parameter--this is why the tests were being
discovered but were not being run
- The `TestClass`, `TestMethod`, `ClassInitialize`, and `ClassCleanup`
decorators aren't intended to have parentheses. They still work with the
parentheses, but it is a nonstandard representation. I fixed these
wherever I found them (4 test classes)

##### Motivation

Test coverage and consistency

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue
- [n/a] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.